+++ src/rTree.ts
@@ -136,11 +136,18 @@ export default class implements vscode.TreeDataProvider<Record> {
 
 export class Record extends vscode.TreeItem {
     constructor(rdoc: utils.RecordDoc, pdoc: utils.ProblemDoc | string, udoc: utils.UserDoc | string) {
+        if (typeof pdoc !== 'string') {
+            pdoc = pdoc.title;
+        }
+        if (typeof udoc !== 'string') {
+            udoc = udoc.uname;
+        }
+
         super(rdoc.score + ' ' + utils.statusName[rdoc.status], vscode.TreeItemCollapsibleState.None);
         this.id = rdoc._id;
         this.contextValue = 'record';
         if (rdoc.pid) {
-            this.description = 'P' + rdoc.pid + ' ' + (typeof pdoc === 'string' ? pdoc : pdoc.title);
+            this.description = `P${rdoc.pid} ${pdoc}  ${udoc}`;
         }
         else {
             this.description = '*';
@@ -148,7 +155,7 @@ export class Record extends vscode.TreeItem {
         this.iconPath = path.join(__dirname, '..', 'res', 'icons', utils.statusIcon[rdoc.status] + '.svg');
         const TooltipDoc = new vscode.MarkdownString();
         TooltipDoc.appendMarkdown(`- **Status**: ${utils.statusName[rdoc.status]}\n`);
-        TooltipDoc.appendMarkdown(`- **User**: ${typeof udoc === 'string' ? udoc : udoc.uname}\n`);
+        TooltipDoc.appendMarkdown(`- **User**: ${udoc}\n`);
         TooltipDoc.appendMarkdown(`- **Score**: ${rdoc.score}\n`);
         if (rdoc.time) { TooltipDoc.appendMarkdown(`- **Time**: ${utils.toTime(rdoc.time)}\n`); }
         if (rdoc.memory) { TooltipDoc.appendMarkdown(`- **Memory**: ${utils.toMemory(rdoc.memory)}\n`); }