+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.24
 
+- Add username in record tree view
 - Fix the tried count
 - Close the problem after open in problem set
 - Fix code and `Open in Problem Set` button